+++ Pinta.Core/Managers/EffectsManager.cs
@@ -73,7 +73,7 @@ namespace Pinta.Core
 			else
 				menu_item = act.CreateAcceleratedMenuItem (adjustment.AdjustmentMenuKey, adjustment.AdjustmentMenuKeyModifiers);
 
-			((Menu)((ImageMenuItem)PintaCore.Chrome.MainMenu.Children[5]).Submenu).AppendMenuItemSorted (menu_item);
+			((Menu)((ImageMenuItem)PintaCore.Chrome.MainMenu.Children[1]).Submenu).AppendMenuItemSorted (menu_item);
 
 			adjustments.Add (adjustment, act);
 			adjustment_menuitems.Add (adjustment, menu_item);